chore: reuse modal
